Update sl ansible interface from manual testing
This commit updates some of the pieces involved with launching and
deleting sl vms via ansible. I've manually tested the sl_launch.yml
playbook works as expected and the VM deletion at least results in a
ticket to cancel the vm being opened. (although I can use the vm) Also
the docstring for parameters are expanded, because the requests are
fairly particular about the formatting.