Add the spark deploy stage to deploy_cluster()
This commit adds the steps needed to the deploy_cluster() function to
actually deploy the spark cluster on all the nodes. When it is called
it will first create the nodes, then setup python on them, followed by
installing spark. With this it should hopefully be easy to now run this
from with in dsx.