Pivot to use centos instead of ubuntu
It turns out the playbook and roles we're using for deploying spark all
assume the use of yum and not apt-get. So to make everything actually
work we need to use a distro that actually does that. This commit
switches all the playbooks to use centos instead of ubuntu so we can
actually deploy spark.