Make server launch multithreaded
SoftLayer launching virtual servers is quite slow, and can sometimes
take 10-20mins to bring a VM online for use. Since we need to wait for
all the servers to become available before we can deploy the cluster
doing this serially can take a really long time, especially for larger
clusters. This commit makes the actual launching of the servers phase
multi-threaded. A thread is started for each server being launched and
they are all started at the same time. Then we wait for them all to
finish before moving on to the deploy stages.