Only deploy a cluster once
We don't want to deploy a cluster more than once. That would be a race
of resources, and potentially confusing fatal for parts of the deploy
code which assume a single server per hostname. This commit just adds a
check if the server deploy stage has already happened or not.