Actually start worker threads
In an earlier commit we made things multi-threaded, but that missed a
critical component, actually starting the thread for each VM launch.
This resulted in an error when trying to join on the threads, because
nothing was actually started. This commit fixes that by starting each
thread right after it's created.