Fix launcher Thread arguments
The args parameter for a thread needs to be an interable. Previously we
intended to pass a tuple with a single int, but because the comma was
forgotten we were passing a single int. This is not iterable and the
function failed. This commit fixes that by adding the missing comma and
passing a tuple with a single entry.