Write inventory file after deploy
Since moving to a multithreaded async launch process the inventory file
writing has been broken. It was originally written assuming a fixed
order, but now that the workers can finish launching in any order there
is no guarantee that the output will be in the right order. This commit
fixes this by moving the inventory file construction to after all the
nodes finish launching.