Remove loopback references to hostname
By default centos is adding an entry to /etc/hosts with the hostname for
127.0.0.1 and ::1. Which makes normally makes sense, you want local
connections to go on the loopback interface instead of going out and
back over a real interface. However, in the case of the spark cluster,
spark is being a bit too clever and detecting the hostname resolves to a
loopback interface and helpfully picking the first network interface. In
the case of the softlayer VMs this is eth0, or the backend ip. This
causes all the slaves to be unable to reach the master because it's not
listening on the public interface they're using. Since everything is
being done via hostnames and it appears the backend interfaces need
additional work to be able to talk to each other we can't just update
the hosts file to use the internal ips instead of public. This patch is
a quick fix to workaround this issue, it just removes the loopback
references from the hosts file. This will cause the spark to realize
that it's listening on the public interface and bind to it. Then the
slaves should be able to talk to the master.