Fix missing param from start_spark.yml playbook
The previous commit added a couple new tasks to remove the loopback
entry for the hostname from /etc/hosts. But this patch neglected to add
the 'with_items:' parameter to tell ansible which hosts to run those
tasks over. This commit fixes that oversight and adds the missing
parameter to tell ansible to run it all hosts in the inventory file.