Add missing python module
An earlier commit claimed it fixed an issue generating inventory files
by adding a new module to deterministically write the output. However,
that commit neglected to actually add the file, it just included an
import in the dsxspark.launch_cluster module that referenced it. This
commit fixes this grave oversight and actually adds the file.