Add playbook to start livy service
There are some scenarios where you need livy running for client apps to
talk to the spark cluster. This commit adds a playbook which will start
the process for doing that. This still isn't tested or used anywhere yet
and it's not clear if this is needed. But just in case this adds it. If
it turns out we don't need it, it'll be removed in the future.