Add support for OpenStack clouds
This commit adds support for deploying the spark cluster to an OpenStack
cloud instead of softlayer's vm service. Right now this works fine, but
depending on the image the ansible playbooks for installing spark might
not work if the ssh user isn't root. Doc updates will come in a follow
up patch.