Add files and symlinks to enable non-root usage
The external spark install module is hard coded assuming it's being run
as root. However there are some usecases where remote access via root is
disabled. To work around that we need to privlege escalate, but since
the roles and playbook for installing spark are outside of the dsxspark
project we need to play some games to write our own playbook and use the
external roles.