Setup things to run playbooks on slow clouds with non-root users
This commit finishes the setup for running things on my private cloud.
There were some last minute fixes needed for getting things running as
non-root and using sudo for things.

Also, the centos image being used is fairly slow and requires a timeout
bump, this is safe to do for all envs, because it only will delay failure
modes.